+++ src/components/AddCardPage/AddCardForm/AddCardInput/CVCInput/styles.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+import Container from '../../../../shared/Container';
+
+export const QuestionMarkContainer = styled(Container)`
+  width: fit-content;
+  height: fit-content;
+  position: relative;
+  display: flex;
+  align-items: center;
+  margin: auto 0 auto 1rem;
+  cursor: pointer;
+
+  &:hover::after {
+    content: '';
+    position: absolute;
+    display: block;
+    width: 10rem;
+    height: 10rem;
+    left: 2.5rem;
+    top: -2.5rem;
+    background: url(${process.env.PUBLIC_URL + '/images/example.png'}) no-repeat;
+  }
+`;
