+++ src/components/AddCardPage/AddCardForm/AddCardInputContainer/index.tsx
@@ -4,10 +4,17 @@ import PALETTE from '../../../../constants/palette';
 
 interface Props {
   width?: string;
+  backgroundColor?: string;
 }
 
-export const AddCardInputContainer: FC<Props> = ({ children, width }) => (
-  <Container flex justifyContent="center" alignItems="center" backgroundColor={PALETTE.GRAY_1} width={width}>
+export const AddCardInputContainer: FC<Props> = ({ children, width, backgroundColor }) => (
+  <Container
+    flex
+    justifyContent="center"
+    alignItems="center"
+    backgroundColor={backgroundColor || PALETTE.GRAY_1}
+    width={width}
+  >
     {children}
   </Container>
 );
