+++ src/components/AddCardPage/AddCardForm/index.tsx
@@ -1,34 +1,18 @@
-import { ChangeEvent, FocusEvent, FormEvent, useEffect, useRef, useState } from 'react';
-import { CARD_BRAND, LABEL, PLACEHOLDER } from '../../../constants/addCardForm';
-import {
-  CARD_NUMBER_DIGITS,
-  CARD_NUMBER_SEPARATOR,
-  CVC_DIGITS,
-  EXP_DATE_DIGITS,
-  EXP_DATE_WHITESPACE_CHARACTER,
-  MAX_OWNER_NAME_LENGTH,
-} from '../../../constants/creditCard';
+import { FormEvent, useEffect, useState } from 'react';
+import { CARD_BRAND } from '../../../constants/addCardForm';
+import { CARD_NUMBER_DIGITS, CARD_NUMBER_SEPARATOR } from '../../../constants/creditCard';
 import { ALERT } from '../../../constants/messages';
 import { CardBrand, ExpDate } from '../../../types';
 import Button from '../../shared/Button';
-import Container from '../../shared/Container';
-import CreditCard from '../../shared/CreditCard';
-import QuestionIcon from '../../shared/Icon/QuestionIcon';
-import Input from '../../shared/Input';
 import CardBrandModal from '../modal/CardBrandModal';
 import NicknameModal from '../modal/NicknameModal';
-import AddCardInputLabel from './AddCardInputLabel';
-import AddCardInputContainer from './AddCardInputContainer';
-import { AddCardFormContainer } from './styles';
-import {
-  isAllInputFilled,
-  isValidCardNumber,
-  isValidCVC,
-  isValidExpMonth,
-  isValidExpYear,
-  isValidOwnerName,
-  isValidPassword,
-} from './validator';
+import CardNumberInput from './AddCardInput/CardNumberInput';
+import CVCInput from './AddCardInput/CVCInput';
+import ExpDateInput from './AddCardInput/ExpDateInput';
+import OwnerNameInput from './AddCardInput/OwnerNameInput/index ';
+import PasswordInput from './AddCardInput/PasswordInput';
+import { AddCardFormCreditCard } from './styles';
+import { isAllInputFilled } from './validator';
 
 export type CardNumberState = [string, string, string, string];
 export type PasswordState = [string, string];
@@ -43,78 +27,6 @@ const AddCardForm = () => {
   const [CVC, setCVC] = useState<string>('');
   const [isCardBrandModalVisible, setIsCardBrandModalVisible] = useState<boolean>(false);
   const [isNicknameModalVisible, setIsNicknameModalVisible] = useState<boolean>(false);
-
-  const secondCardNumberInputRef = useRef<HTMLInputElement>(null);
-  const thirdCardNumberInputRef = useRef<HTMLInputElement>(null);
-  const fourthCardNumberInputRef = useRef<HTMLInputElement>(null);
-  const expYearInputRef = useRef<HTMLInputElement>(null);
-  const secondPasswordInputRef = useRef<HTMLInputElement>(null);
-
-  const focusNextCardNumberInput = (index: number) => {
-    const cardNumberInputRefs = [secondCardNumberInputRef, thirdCardNumberInputRef, fourthCardNumberInputRef];
-
-    cardNumberInputRefs[index].current?.focus();
-  };
-
-  const onChangeCardNumber = ({ target: { value } }: ChangeEvent<HTMLInputElement>, index: number) => {
-    if (!isValidCardNumber(value)) return;
-
-    const nextValue: CardNumberState = [...cardNumber];
-
-    nextValue[index] = value;
-    setCardNumber(nextValue);
-
-    if (nextValue[index].length === CARD_NUMBER_DIGITS && index < 3) {
-      focusNextCardNumberInput(index);
-    }
-  };
-
-  const onChangeExpMonth = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-    if (!isValidExpMonth(value)) return;
-
-    setExpDate({ ...expDate, month: value });
-
-    if (value.length === EXP_DATE_DIGITS) {
-      expYearInputRef.current?.focus();
-    }
-  };
-
-  const onChangeExpYear = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-    if (!isValidExpYear(value)) return;
-
-    setExpDate({ ...expDate, year: value });
-  };
-
-  const onBlurExpDate = ({ target: { value } }: FocusEvent<HTMLInputElement>, index: keyof ExpDate) => {
-    if (value.length !== 1) return;
-
-    setExpDate({ ...expDate, [index]: EXP_DATE_WHITESPACE_CHARACTER + value });
-  };
-
-  const onChangeOwnerName = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-    if (!isValidOwnerName(value)) return;
-
-    setOwnerName(value.toUpperCase());
-  };
-
-  const onChangeCVC = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-    if (!isValidCVC(value)) return;
-
-    setCVC(value);
-  };
-
-  const onChangePassword = ({ target: { value } }: ChangeEvent<HTMLInputElement>, index: 0 | 1) => {
-    if (!isValidPassword(value)) return;
-
-    const nextPassword: PasswordState = [...password];
-
-    nextPassword[index as number] = value;
-    setPassword(nextPassword);
-
-    if (index === 0 && value.length === 1) {
-      secondPasswordInputRef.current?.focus();
-    }
-  };
 
   const onSelectCardBrand = (cardBrand: CardBrand) => {
     setCardBrand(cardBrand);
@@ -165,114 +77,21 @@ const AddCardForm = () => {
   };
 
   return (
-    <AddCardFormContainer>
-      <CreditCard
-        className="credit-card"
+    <div>
+      <AddCardFormCreditCard
         cardBrand={cardBrand}
         expDate={expDate}
         ownerName={ownerName}
         cardNumber={cardNumber.join(CARD_NUMBER_SEPARATOR)}
       />
+
       <form onSubmit={onSubmitCard}>
-        <AddCardInputLabel label={LABEL.CARD_NUMBER}>
-          <AddCardInputContainer>
-            {[
-              { type: 'text', ref: null },
-              { type: 'text', ref: secondCardNumberInputRef },
-              { type: 'password', ref: thirdCardNumberInputRef },
-              { type: 'password', ref: fourthCardNumberInputRef },
-            ]
-              .map(({ type, ref }, index) => (
-                <Input
-                  key={index}
-                  type={type}
-                  ref={ref}
-                  textCenter
-                  maxLength={CARD_NUMBER_DIGITS}
-                  width="16%"
-                  value={cardNumber[index]}
-                  onChange={event => onChangeCardNumber(event, index)}
-                />
-              ))
-              .reduce(
-                (acc: JSX.Element[], curr: JSX.Element, index, array: JSX.Element[]) => [
-                  ...acc,
-                  <span key={index + array.length}>{CARD_NUMBER_SEPARATOR}</span>,
-                  curr,
-                ],
-                []
-              )
-              .slice(1)}
-          </AddCardInputContainer>
-        </AddCardInputLabel>
-        <AddCardInputLabel label={LABEL.EXP_DATE} width="40%">
-          <AddCardInputContainer>
-            <Input
-              type="text"
-              placeholder={PLACEHOLDER.MONTH}
-              textCenter
-              maxLength={EXP_DATE_DIGITS}
-              width="40%"
-              value={expDate.month}
-              onChange={onChangeExpMonth}
-              onBlur={event => onBlurExpDate(event, 'month')}
-            />
-            /
-            <Input
-              type="text"
-              ref={expYearInputRef}
-              placeholder={PLACEHOLDER.YEAR}
-              textCenter
-              maxLength={EXP_DATE_DIGITS}
-              width="40%"
-              value={expDate.year}
-              onChange={onChangeExpYear}
-              onBlur={event => onBlurExpDate(event, 'year')}
-            />
-          </AddCardInputContainer>
-        </AddCardInputLabel>
-        <AddCardInputLabel label={[LABEL.OWNER_NAME, `${ownerName.length} / ${MAX_OWNER_NAME_LENGTH}`]}>
-          <AddCardInputContainer>
-            <Input
-              placeholder={PLACEHOLDER.OWNER_NAME}
-              type="text"
-              width="90%"
-              maxLength={MAX_OWNER_NAME_LENGTH}
-              value={ownerName}
-              onChange={onChangeOwnerName}
-            />
-          </AddCardInputContainer>
-        </AddCardInputLabel>
-        <AddCardInputLabel label={LABEL.CVC}>
-          <AddCardInputContainer width="25%">
-            <Input type="password" maxLength={CVC_DIGITS} textCenter value={CVC} onChange={onChangeCVC} />
-          </AddCardInputContainer>
-          <Container className="question-mark">
-            <QuestionIcon />
-          </Container>
-        </AddCardInputLabel>
-        <AddCardInputLabel label={LABEL.PASSWORD}>
-          <Container flex justifyContent="space-between" width="60%">
-            <AddCardInputContainer width="23%">
-              <Input type="password" textCenter value={password[0]} onChange={event => onChangePassword(event, 0)} />
-            </AddCardInputContainer>
-            <AddCardInputContainer width="23%">
-              <Input
-                type="password"
-                ref={secondPasswordInputRef}
-                textCenter
-                value={password[1]}
-                onChange={event => onChangePassword(event, 1)}
-              />
-            </AddCardInputContainer>
-            <AddCardInputContainer width="23%">
-              <span className="password-dot" />
-            </AddCardInputContainer>
-            <AddCardInputContainer width="23%">
-              <span className="password-dot" />
-            </AddCardInputContainer>
-          </Container>
-        </AddCardInputLabel>
+        <CardNumberInput cardNumber={cardNumber} setCardNumber={setCardNumber} />
+        <ExpDateInput expDate={expDate} setExpDate={setExpDate} />
+        <OwnerNameInput ownerName={ownerName} setOwnerName={setOwnerName} />
+        <CVCInput CVC={CVC} setCVC={setCVC} />
+        <PasswordInput password={password} setPassword={setPassword} />
+
         <Button type="button" position="bottom-right" onClick={onClickNextButton}>
           다음
         </Button>
@@ -296,7 +115,7 @@ const AddCardForm = () => {
           />
         )}
       </form>
-    </AddCardFormContainer>
+    </div>
   );
 };
